requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.21.0
PyYAML>=6.0
pyarrow>=10.0.0  # For Parquet file support

# Development and Testing (optional)
jupyter>=1.0.0
ipykernel>=6.0.0

# Type hints support
typing-extensions>=4.0.0